+++ sink_writetofile.go
@@ -40,7 +40,6 @@
 package scriptish
 
 import (
-	"io"
 	"os"
 )
 
@@ -62,11 +61,31 @@ func WriteToFile(filename string) Command {
 		defer fh.Close()
 
 		// write to the file
-		_, err = io.Copy(fh, p.Stdin)
-		if err != nil {
-			return StatusNotOkay, err
+		if p.Flags&contextIsPipeline != 0 {
+			// we are in a pipeline
+			for line := range p.Stdin.ReadLines() {
+				_, err = fh.WriteString(line)
+				if err != nil {
+					return StatusNotOkay, err
+				}
+				_, err = fh.WriteString("\n")
+				if err != nil {
+					return StatusNotOkay, err
+				}
+			}
+		} else {
+			// we are in a list
+			for line := range p.Stdout.ReadLines() {
+				_, err = fh.WriteString(line)
+				if err != nil {
+					return StatusNotOkay, err
+				}
+				_, err = fh.WriteString("\n")
+				if err != nil {
+					return StatusNotOkay, err
+				}
+			}
 		}
-
 		// all done
 		return StatusOkay, err
 	}